packages/client/pyproject.toml
[build-system]
requires = ["editables>=0.5", "hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "xrpl-mpp-client"
version = "0.1.3"
description = "Buyer-side Python SDK for XRPL MPP HTTP payments"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.12"
license = "MIT"
authors = [
  { name = "Louis-Guillaume Carrier-Bedard", email = "lgcarrier@gmail.com" },
]
keywords = ["xrpl", "mpp", "client", "httpx", "payments"]
classifiers = [
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Topic :: Internet :: WWW/HTTP",
  "Typing :: Typed",
]
dependencies = [
  "httpx==0.28.1",
  "pydantic>=2,<3",
  "xrpl-py==4.5.0",
  "xrpl-mpp-core>=0.1.3,<0.2.0",
]

[project.urls]
Homepage = "https://github.com/lgcarrier/xrpl-mpp-stack"
Documentation = "https://lgcarrier.github.io/xrpl-mpp-stack/packages/client/"
Repository = "https://github.com/lgcarrier/xrpl-mpp-stack"
Issues = "https://github.com/lgcarrier/xrpl-mpp-stack/issues"
Changelog = "https://github.com/lgcarrier/xrpl-mpp-stack/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/xrpl_mpp_client"]

[tool.hatch.build.hooks.custom]
path = "build_hook.py"

[tool.hatch.build.targets.wheel.force-include]
"LICENSE" = "LICENSE"

[tool.hatch.build.targets.sdist]
include = [
  "/src/xrpl_mpp_client",
  "/README.md",
  "/LICENSE",
  "/build_hook.py",
]
